# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Exam Arena API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface ResendVerificationInput {
  email: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
  /** @minLength 1 */
  name: string;
  /**
     * Optional referral code from an existing user
     * @nullable
     */
  referralCode?: string | null;
}

export interface LoginInput {
  email: string;
  password: string;
  rememberMe?: boolean;
}

export type UserSubscriptionPlanProperty = typeof UserSubscriptionPlanProperty[keyof typeof UserSubscriptionPlanProperty];


export const UserSubscriptionPlanProperty = {
  none: 'none',
  trial: 'trial',
  monthly: 'monthly',
  semester: 'semester',
  yearly: 'yearly',
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  subscriptionPlan: UserSubscriptionPlanProperty;
  /** @nullable */
  subscriptionEnd?: string | null;
  /** @nullable */
  semesterStart?: string | null;
  /** @nullable */
  referralCode?: string | null;
  /** @nullable */
  momoNumber?: string | null;
  /** @nullable */
  momoName?: string | null;
}

export interface AuthResponse {
  user: User;
}

export interface ForgotPasswordInput {
  email: string;
}

export interface ResetPasswordInput {
  token: string;
  /** @minLength 6 */
  password: string;
}

export interface Question {
  id: number;
  year: string;
  subject: string;
  week: number;
  weekTopic: string;
  questionNumber: number;
  questionText: string;
  optionA: string;
  optionB: string;
  optionC: string;
  optionD: string;
  correctAnswer: string;
  /** Images or diagrams positioned with this question in the uploaded document */
  questionImages: string[];
  /** @nullable */
  dok?: string | null;
  /** @nullable */
  learningIndicator?: string | null;
  /** @nullable */
  feedback?: string | null;
}

export interface QuestionFilters {
  years: string[];
  subjects: string[];
  weeks: number[];
}

export interface QuizSessionInput {
  year: string;
  subject: string;
  week: number;
}

export interface QuizSession {
  id: number;
  year: string;
  subject: string;
  week: number;
  /** @nullable */
  weekTopic?: string | null;
  /** @nullable */
  score?: number | null;
  totalQuestions: number;
  /** @nullable */
  completedAt: string | null;
}

export interface QuizAnswer {
  questionId: number;
  /** @nullable */
  selectedAnswer?: string | null;
  isCorrect: boolean;
}

export interface QuizSessionDetail {
  id: number;
  year: string;
  subject: string;
  week: number;
  /** @nullable */
  weekTopic?: string | null;
  /** @nullable */
  score?: number | null;
  totalQuestions?: number;
  /** @nullable */
  completedAt?: string | null;
  questions: Question[];
  answers: QuizAnswer[];
}

export interface AnswerInput {
  questionId: number;
  selectedAnswer: string;
}

export interface QuizSubmitInput {
  answers: AnswerInput[];
}

export interface Song {
  id: number;
  title: string;
  url: string;
  sortOrder: number;
  isActive: boolean;
}

export interface SongUploadInput {
  title: string;
}

export interface SongUpdate {
  title?: string;
  isActive?: boolean;
  sortOrder?: number;
}

export interface SongReorderInput {
  songIds: number[];
}

export type MomoPaymentInputPlan = typeof MomoPaymentInputPlan[keyof typeof MomoPaymentInputPlan];


export const MomoPaymentInputPlan = {
  monthly: 'monthly',
  semester: 'semester',
  yearly: 'yearly',
} as const;

export interface MomoPaymentInput {
  plan: MomoPaymentInputPlan;
  /** @minLength 1 */
  txId: string;
  /** @nullable */
  semesterStart?: string | null;
}

export interface MomoPaymentResponse {
  message: string;
  reference: string;
}

export interface SubscriptionStatus {
  plan: string;
  isActive: boolean;
  /** @nullable */
  subscriptionEnd?: string | null;
  /** @nullable */
  daysRemaining?: number | null;
}

export interface MomoDetailsInput {
  /**
     * MoMo phone number (e.g. 0241234567)
     * @minLength 10
     */
  momoNumber: string;
  /**
     * Name registered on MoMo — must match exactly
     * @minLength 1
     */
  momoName: string;
}

export interface ReferralEarningEntry {
  id: number;
  refereeName: string;
  plan: string;
  amount: number;
  status: string;
  createdAt: string;
}

export interface ReferralInfo {
  referralCode: string;
  totalEarningsPesewas: number;
  pendingEarningsPesewas: number;
  /** @nullable */
  momoName: string | null;
  /** @nullable */
  momoNumber: string | null;
  earnings: ReferralEarningEntry[];
}

export interface AdminReferralRow {
  userId: number;
  userName: string;
  userEmail: string;
  /** @nullable */
  momoName: string | null;
  /** @nullable */
  momoNumber: string | null;
  pendingAmount: number;
  paidAmount: number;
  earnings: ReferralEarningEntry[];
}

export interface AdminTestEmailInput {
  to?: string;
}

export interface AdminLoginInput {
  email?: string;
  password: string;
}

export interface QuestionUploadInput {
  year?: string;
  subject?: string;
}

export interface UploadResult {
  inserted: number;
  skipped: number;
  errors?: string[];
  preview?: Question[];
}

export interface UserAdmin {
  id: number;
  email: string;
  name: string;
  subscriptionPlan: string;
  /** @nullable */
  subscriptionEnd?: string | null;
  createdAt: string;
}

export interface AdminStats {
  totalUsers: number;
  activeSubscribers: number;
  totalQuestions: number;
  totalSubjects: number;
  totalSongs: number;
  recentSessions: number;
}

export interface ActivityLog {
  id: number;
  type: string;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  userEmail?: string | null;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  metadata?: string | null;
  /** @nullable */
  ip?: string | null;
  createdAt: string;
}

export interface AdminPayment {
  id: number;
  userId: number;
  /** @nullable */
  userEmail?: string | null;
  /** @nullable */
  userName?: string | null;
  plan: string;
  amount: number;
  status: string;
  reference: string;
  /** @nullable */
  userTxId?: string | null;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  createdAt: string;
}

export interface MomoVerifyInput {
  txId: string;
}

export interface MomoVerifyResponse {
  match: boolean;
  message: string;
}

export type AdminSubscribeInputPlan = typeof AdminSubscribeInputPlan[keyof typeof AdminSubscribeInputPlan];


export const AdminSubscribeInputPlan = {
  trial: 'trial',
  monthly: 'monthly',
  semester: 'semester',
  yearly: 'yearly',
  lifetime: 'lifetime',
} as const;

export interface AdminSubscribeInput {
  email: string;
  plan: AdminSubscribeInputPlan;
  months?: number;
  generatePassword?: boolean;
}

export interface AdminSubscribeResponse {
  message: string;
  /** @nullable */
  generatedPassword?: string | null;
}

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  name: string;
  /** Average score percentage (0–100, rounded to 1 decimal) */
  avgScore: number;
  /** Number of completed quiz sessions counted */
  quizzesCompleted: number;
}

export type UserSubscriptionPlan = typeof UserSubscriptionPlan[keyof typeof UserSubscriptionPlan];


export const UserSubscriptionPlan = {
  none: 'none',
  trial: 'trial',
  monthly: 'monthly',
  semester: 'semester',
  yearly: 'yearly',
} as const;

export type ListQuestionsParams = {
year?: string;
subject?: string;
week?: number;
};

export type GetQuestionFiltersParams = {
year?: string;
subject?: string;
};

export type DeleteQuestionsByFilterParams = {
year: string;
subject: string;
};

export type ListAdminActivityParams = {
limit?: number;
type?: string;
};