# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Exam Arena API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quiz/sessions:
    post:
      operationId: createQuizSession
      tags: [quiz]
      summary: Start a quiz session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizSessionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizSession"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /quiz/sessions/{sessionId}:
    get:
      operationId: getQuizSession
      tags: [quiz]
      summary: Get quiz session with results
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizSessionDetail"
  /quiz/sessions/{sessionId}/submit:
    post:
      operationId: submitQuizSession
      tags: [quiz]
      summary: Submit quiz answers
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizSubmitInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizSessionDetail"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Question:
      type: object
      required: [id, year, subject, week, weekTopic, questionNumber, questionText, optionA, optionB, optionC, optionD, correctAnswer, questionImages]
      properties:
        id:
          type: integer
        year:
          type: string
        subject:
          type: string
        week:
          type: integer
        weekTopic:
          type: string
        questionNumber:
          type: integer
        questionText:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctAnswer:
          type: string
        questionImages:
          type: array
          description: Images or diagrams positioned with this question in the uploaded document
          items:
            type: string
        dok:
          type: ["string", "null"]
        learningIndicator:
          type: ["string", "null"]
        feedback:
          type: ["string", "null"]
...
    QuizSessionInput:
      type: object
      required: [year, subject, week]
      properties:
        year:
          type: string
        subject:
          type: string
        week:
          type: integer
    QuizSession:
      type: object
      required: [id, year, subject, week, totalQuestions, completedAt]
      properties:
        id:
          type: integer
        year:
          type: string
        subject:
          type: string
        week:
          type: integer
        weekTopic:
          type: ["string", "null"]
        score:
          type: ["integer", "null"]
        totalQuestions:
          type: integer
        completedAt:
          type: ["string", "null"]
    QuizSessionDetail:
      type: object
      required: [id, year, subject, week, questions, answers]
      properties:
        id:
          type: integer
        year:
          type: string
        subject:
          type: string
        week:
          type: integer
        weekTopic:
          type: ["string", "null"]
        score:
          type: ["integer", "null"]
        totalQuestions:
          type: integer
        completedAt:
          type: ["string", "null"]
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        answers:
          type: array
          items:
            $ref: "#/components/schemas/QuizAnswer"
    QuizAnswer:
      type: object
      required: [questionId, isCorrect]
      properties:
        questionId:
          type: integer
        selectedAnswer:
          type: ["string", "null"]
        isCorrect:
          type: boolean
    QuizSubmitInput:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/AnswerInput"
    AnswerInput:
      type: object
      required: [questionId, selectedAnswer]
      properties:
        questionId:
          type: integer
        selectedAnswer:
          type: string

    # ── Songs ─────────────────────────────────────────────────────────────────
...