# lib/api-zod/src/generated/types/question.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Exam Arena API
 * OpenAPI spec version: 0.1.0
 */

export interface Question {
  id: number;
  year: string;
  subject: string;
  week: number;
  weekTopic: string;
  questionNumber: number;
  questionText: string;
  optionA: string;
  optionB: string;
  optionC: string;
  optionD: string;
  correctAnswer: string;
  /** Images or diagrams positioned with this question in the uploaded document */
  questionImages: string[];
  /** @nullable */
  dok?: string | null;
  /** @nullable */
  learningIndicator?: string | null;
  /** @nullable */
  feedback?: string | null;
}